requirements.txt
Django
djangorestframework
Pillow
django-ckeditor
django-cors-headers
django-rest-auth